Added 'Select All' button to Searched Items
Added a 'select all' button to the Searched Items view in the course
search settings. This makes it easy to select or deslect all of the options.
The button will also update when the user manually selects all of the options
one at a time.

This also contains formatting changes to the code in the component file.